UserController updated.
